feat: adjusted based on response
Split Beamswap folder into Beamex and Beamswap-v3
