+++ dexs/beamswap/index.ts
@@ -1,101 +1,15 @@
-import { gql, request } from "graphql-request";
-import { BreakdownAdapter, ChainEndpoints, Fetch } from "../../adapters/types";
+import { BreakdownAdapter, ChainEndpoints } from "../../adapters/types";
 import { CHAIN } from "../../helpers/chains";
 import customBackfill from "../../helpers/customBackfill";
 import { getStartTimestamp } from "../../helpers/getStartTimestamp";
-import {
-  DEFAULT_DAILY_VOLUME_FACTORY,
-  DEFAULT_TOTAL_VOLUME_FIELD,
-  getGraphDimensions,
-} from "../../helpers/getUniSubgraph";
-import { getUniqStartOfTodayTimestamp } from "../../helpers/getUniSubgraphVolume";
+import { getGraphDimensions } from "../../helpers/getUniSubgraph";
 
 const endpoints: ChainEndpoints = {
   [CHAIN.MOONBEAN]:
     "https://api.thegraph.com/subgraphs/name/beamswap/beamswap-dex-v2",
 };
 
-const endpointsBeamex: ChainEndpoints = {
-  [CHAIN.MOONBEAN]:
-    "https://api.thegraph.com/subgraphs/name/flisko/stats-moonbeam",
-};
-const historicalDataSwap = gql`
-  query get_volume($period: String!, $id: String!) {
-    volumeStats(where: { period: $period, id: $id }) {
-      swap
-    }
-  }
-`;
 
-const historicalDataDerivatives = gql`
-  query get_volume($period: String!, $id: String!) {
-    volumeStats(where: { period: $period, id: $id }) {
-      liquidation
-      margin
-    }
-  }
-`;
-
-interface IGraphResponse {
-  volumeStats: Array<{
-    burn: string;
-    liquidation: string;
-    margin: string;
-    mint: string;
-    swap: string;
-  }>;
-}
-
-const getFetch =
-  (query: string) =>
-  (chain: string): Fetch =>
-  async (timestamp: number) => {
-    const dayTimestamp = getUniqStartOfTodayTimestamp(
-      new Date(timestamp * 1000)
-    );
-    const dailyData: IGraphResponse = await request(
-      endpointsBeamex[chain],
-      query,
-      {
-        id: String(dayTimestamp),
-        period: "daily",
-      }
-    );
-    const totalData: IGraphResponse = await request(
-      endpointsBeamex[chain],
-      query,
-      {
-        id: "total",
-        period: "total",
-      }
-    );
-
-    return {
-      timestamp: dayTimestamp,
-      dailyVolume:
-        dailyData.volumeStats.length == 1
-          ? String(
-              Number(
-                Object.values(dailyData.volumeStats[0]).reduce((sum, element) =>
-                  String(Number(sum) + Number(element))
-                )
-              ) *
-                10 ** -30
-            )
-          : undefined,
-      totalVolume:
-        totalData.volumeStats.length == 1
-          ? String(
-              Number(
-                Object.values(totalData.volumeStats[0]).reduce((sum, element) =>
-                  String(Number(sum) + Number(element))
-                )
-              ) *
-                10 ** -30
-            )
-          : undefined,
-    };
-  };
 
 const graphs = getGraphDimensions({
   graphUrls: endpoints,
@@ -143,30 +57,6 @@ const v1graphs = getGraphDimensions({
   },
 });
 
-const endpointV3 = {
-  [CHAIN.MOONBEAM]:
-    "https://api.thegraph.com/subgraphs/name/beamswap/beamswap-v3",
-};
-const VOLUME_USD = "volumeUSD";
-const v3Graphs = getGraphDimensions({
-  graphUrls: endpointV3,
-  totalVolume: {
-    factory: "factories",
-    field: DEFAULT_TOTAL_VOLUME_FIELD,
-  },
-  dailyVolume: {
-    factory: DEFAULT_DAILY_VOLUME_FACTORY,
-    field: VOLUME_USD,
-  },
-  feesPercent: {
-    type: "fees",
-    ProtocolRevenue: 16,
-    HoldersRevenue: 0,
-    UserFees: 100, // User fees are 100% of collected fees
-    SupplySideRevenue: 84, // 84% of fees are going to LPs
-    Revenue: 0,
-  },
-});
 
 const methodology = {
   UserFees: "User pays 0.30% fees on each swap.",
@@ -177,12 +67,6 @@ const methodology = {
   HoldersRevenue: "Stakers received $GLINT in staking rewards.",
 };
 
-const methodologyv3 = {
-  UserFees: "User pays 0.01%, 0.05%, 0.3%, or 1% on each swap.",
-  ProtocolRevenue: "Protocol receives 16% of fees.",
-  SupplySideRevenue: "84% of user fees are distributed among LPs.",
-  HoldersRevenue: "Holders have no revenue.",
-};
 const methodologyStable = {
   UserFees: "User pays a 0.04% fee on each swap.",
   Fees: "A 0.04% of each swap is collected as trading fees",
@@ -192,17 +76,6 @@ const methodologyStable = {
   HoldersRevenue: "Stakers received $GLINT in staking rewards.",
 };
 
-const methodologyBeamex = {
-  Fees: "Fees from open/close position (0.2%), liquidations, swap (0.2% to 0.4%), mint and burn (based on tokens balance in the pool) and borrow fee ((assets borrowed)/(total assets in pool)*0.02%)",
-  UserFees:
-    "Fees from open/close position (0.2%), swap (0.2% to 0.4%) and borrow fee ((assets borrowed)/(total assets in pool)*0.04%)",
-  HoldersRevenue:
-    "30% of all collected fees are distributed to $stGLINT stakers",
-  SupplySideRevenue:
-    "70% of all collected fees will be distributed to BLP stakers. Currently they are distributed to treasury",
-  Revenue: "70% of all collected fees are distributed to the treasury",
-  ProtocolRevenue: "70% of all collected fees are distributed to the treasury",
-};
 const adapter: BreakdownAdapter = {
   breakdown: {
     classic: {
@@ -230,40 +103,6 @@ const adapter: BreakdownAdapter = {
         meta: {
           methodology: {
             ...methodologyStable,
-          },
-        },
-      },
-    },
-    v3: {
-      [CHAIN.MOONBEAN]: {
-        fetch: v3Graphs(CHAIN.MOONBEAN),
-        start: async () => 1684397388,
-        customBackfill: customBackfill(CHAIN.MOONBEAN, v3Graphs),
-        meta: {
-          methodology: {
-            ...methodologyv3,
-          },
-        },
-      },
-    },
-    "beamex-swap": {
-      [CHAIN.MOONBEAN]: {
-        fetch: getFetch(historicalDataSwap)(CHAIN.MOONBEAN),
-        start: async () => 1687421388,
-        meta: {
-          methodology: {
-            ...methodologyBeamex,
-          },
-        },
-      },
-    },
-    "beamex-perps": {
-      [CHAIN.MOONBEAN]: {
-        fetch: getFetch(historicalDataDerivatives)(CHAIN.MOONBEAN),
-        start: async () => 1687421388,
-        meta: {
-          methodology: {
-            ...methodologyBeamex,
           },
         },
       },
